Move unescape function into sanitize, and unescape titles in search results.
